rename MathboxNodeContext -> MathboxAPIContext
